Task1

+++ Tyuiu.SafonovRV.Sprint5.Task1.V1.Lib/DataService.cs
@@ -4,38 +4,40 @@ namespace Tyuiu.SafonovRV.Sprint5.Task1.V1.Lib
 {
     public class DataService : ISprint5Task1V1
     {
-        
-            public string SaveToFileTextData(int startValue, int stopValue)
+        public string SaveToFileTextData(int startValue, int stopValue)
+        {
+            string path = Path.Combine(new string[] { Path.GetTempPath(), "OutPutFileTask1.txt" });
+
+            FileInfo fileInfo = new FileInfo(path);
+            bool fileExists = fileInfo.Exists;
+
+            if (fileExists) { File.Delete(path); }
+
+            double res;
+            string strRes;
+            for (int x = startValue; x <= stopValue; x++)
             {
-                //string path = $@"{Path.Combine("OutPutFileTask1.txt")}";
-                string path = $@"{Path.GetTempPath()}\OutPutFileTask1.txt";
+                res = Math.Round((5 * x + 2.5) / (Math.Sin(x) + 2 + x) + 2, 2);
+                strRes = Convert.ToString(res);
 
-                FileInfo fileInfo = new FileInfo(path);
-                bool fileExists = fileInfo.Exists;
-
-                if (fileExists)
+                if (Math.Sin(x) - 3 + x == 0)
                 {
-                    File.Delete(path);
+                    File.AppendAllText(path, "0");
                 }
 
-                double result;
-                string strResult;
-
-                for (double x = startValue; x <= stopValue; x++)
+                if (x != stopValue)
                 {
-                    result = Math.Round((5*(x))+2.5 / (Math.Sin(x)+3))+2*x+Math.Cos(x);
-                    strResult = Convert.ToString(result);
-
-                    if (x != stopValue)
-                    {
-                        File.AppendAllText(path, strResult + Environment.NewLine);
-                    }
-                    else
-                    {
-                        File.AppendAllText(path, strResult);
-                    }
+                    File.AppendAllText(path, strRes + Environment.NewLine);
                 }
-                return path;
+                else
+                {
+                    File.AppendAllText(path, strRes);
+                }
             }
+            return path;
         }
     }
+}
+
+
+
